requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.0

# LangChain & LangGraph
langchain>=0.1.0
langgraph>=0.0.15
langchain-openai>=0.0.2

# LLM APIs
openai>=1.5.0
google-generativeai>=0.3.0  # For TxGemma

# Data processing and analysis
pandas>=2.0.3
numpy>=1.24.3
scipy>=1.11.3

# Market and Investment Data APIs
yfinance>=0.2.31
alpha_vantage>=2.3.1
requests>=2.31.0
aiohttp>=3.8.6

# Utilities
tqdm>=4.66.1
loguru>=0.7.2
uuid>=1.30
python-dateutil>=2.8.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development utilities
black>=23.10.1
isort>=5.12.0
mypy>=1.6.1

# Core API Dependencies
fastapi==0.95.1
uvicorn==0.22.0
gunicorn==20.1.0

# Additional dependencies you might need later
pydantic==1.10.7
openai==0.27.4